+++ src/css/CoinChart.css
@@ -3,3 +3,19 @@
     width: 100%;   
     border: 1px solid gray;
 }
+
+.select_options {
+    display: flex;
+    margin: 10px;
+    justify-content: flex-start;
+    align-items: center;
+}
+
+.select_options >* {
+    margin-right: 10px;
+    height: 25px;
+}
+
+.datetime, .chart {
+    height: 25px;
+}
